many bugfix
- add the ability to press the button on bootup to erase flash
- give up on multicore again
- fix some flash buffer overflow bugs & memory alignment errors
- add 2axis control mode for planes without rudders (who would have one of those??)
- increase serial timeout because it was causing issues w/ longer commands
- fix loop counter not being incremented in fusion
- do some calibration tests (still yet to come up with a good procedure..)
- fix printed digits in api commands
- fix wifly parsing
- fix implicit function isage
- fix misc api command bugs

+++ src/lib/fusion/fusion.h
@@ -437,7 +437,7 @@ typedef struct SensorFusionGlobals
 	struct PhysicalSensor *pSensors;    	        ///< a linked list of physical sensors
 	volatile uint8_t iPerturbation;	        ///< test perturbation to be applied
 	// Book-keeping variables
-	int32_t loopcounter;			///< counter incrementing each iteration of sensor fusion (typically 25Hz)
+	uint64_t loopcounter;			///< counter incrementing each iteration of sensor fusion (typically 25Hz)
 	int32_t systick_I2C;			///< systick counter to benchmark I2C reads
 	int32_t systick_Spare;			///< systick counter for counts spare waiting for timing interrupt
         ///@}
